Smileyadd deprecated header moved
git-svn-id: http://svn.miranda-ng.org/main/trunk@9440 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SmileyAdd/src/smileyroutines.cpp
@@ -67,7 +67,7 @@ void LookupAllSmileys(SmileyPackType* smileyPack, SmileyPackCType* smileyCPack, 
 	memset(csmlit, 0, smlsz * sizeof(int));
 
 	long numCharsSoFar = 0;
-	int smloff = 0;
+	size_t smloff = 0;
 
 	while (true) {
 		int firstSml = -1;
